+++ Changes.md
@@ -1,9 +1,15 @@
 ## WIP  TBD
 
  * :computer: The `today ost index` command has been added to pull down scripture indexes from openscripture.today.
+ * :computer: Added `--exclude` and `--exclude-index` options to `today random`.
  * ESV text from VerseText and others won't include the references now.
  * Updating the Verse and Photo file formats used by the `ost.Client` for the upcoming version of openscripture.today. Adding a `ost.Metadata` structure that records version, original verseion, and whether the verse and photo have been pruned.
  * A new `VersesIndex` has been added to the `ost.Client` for fetching indexes listing all the verses that have been postd for certain periods (all time, year, month).
+ * Added a `Clone` method to `ref.Canon` and `ref.Book` to create deep clones of these objects.
+ * Added the `ref.ExcludeReferences` option for use with the `ref.Random` functions.
+ * Added a `Subtract` method to `ref.Resolved` to return the difference between two resolved references.
+ * Added a `Filtered` method to `ref.Canon` that returns a new canon that has segments references removed.
+ * :boom: Breaking Change :boom:: Now requires Go 1.22.
  * :boom: Breaking Change :boom:: Significant changes have been made to the photos API. These changes include the following:
    - The `photo.Meta` and `photo.Info` structures have been removed/merged/refactored into `photo.Descriptor`.
    - A new interface named `photo.Image` has been added and each `photo.Descriptor` should have at least one, but may have many associated `photo.Image` objects.